added folders to /src, added profile avatar, centered checkbox
