+++ src/components/Landing.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react'
 import { UserSession } from 'blockstack'
-import { appConfig } from './constants'
-import './Landing.css'
+import { appConfig } from '../assets/constants'
+import '../styles/Landing.css'
 
 
 class Landing extends Component {
