+++ src/components/NavBar.css
@@ -1,3 +1,7 @@
 .navbar {
   background-color:  rgb(33, 31, 109) !important;
 }
+
+.avatar {
+  border-radius:3px
+}
